refactor(messaging): extract the message form as MessageComposer
Pure move, no behaviour change: the send page's form comes out as a
component of its own so a second page can render the same thing.

The two callbacks are what the extraction is for. Peru was a link to the
list and marking a send reported "Lähetetty" in place; both are now
props, so a caller that works through several discs can do something
else with them. SendMessagePage keeps the single-disc behaviour by
passing a cancel that navigates home and no onRecorded.
